% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyGenePlot.R
\name{tidy.FeaturePlot}
\alias{tidy.FeaturePlot}
\title{Title}
\usage{
tidy.FeaturePlot(
  seu_obj,
  features,
  pt.size = 0.5,
  ncol = 1,
  cols = c("gray", "blue")
)
}
\arguments{
\item{seu_obj}{seurat object created with R package Seurat}

\item{features}{a vector containing genes}

\item{pt.size}{point size used for plotting}

\item{ncol}{arrange plots numbers in each row}

\item{cols}{colors used for plotting}
}
\value{
combined_plot
}
\description{
Title
}
