% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyGenePlot.R
\name{tidy.VlnPlot}
\alias{tidy.VlnPlot}
\title{Title}
\usage{
tidy.VlnPlot(seu_obj, features, pt.size = 0.5, ncol = 1, cols = NULL)
}
\arguments{
\item{seu_obj}{seurat object created with R package Seurat}

\item{features}{a vector containing genes}

\item{pt.size}{point size used for plotting}

\item{ncol}{arrange plots numbers in each row}

\item{cols}{colors used for plotting}
}
\value{
combined_plot
}
\description{
Title
}
