% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyGenePlot.R
\name{tidy.VlnPlot}
\alias{tidy.VlnPlot}
\title{Title Generate tidy/clean violin plot(s) from scRNA-seq analysis}
\usage{
tidy.VlnPlot(
  seu_obj,
  features,
  pt.size = 0.5,
  ncol = NULL,
  cols = NULL,
  split.by = NULL
)
}
\arguments{
\item{seu_obj}{seurat object created with R package Seurat}

\item{features}{a vector containing genes}

\item{pt.size}{point size used for plotting}

\item{ncol}{arrange plots numbers in each row}

\item{cols}{colors used for plotting}
}
\value{
combined_plot
}
\description{
Title Generate tidy/clean violin plot(s) from scRNA-seq analysis
}
\examples{
# Examples will be added in a future release.
}
